+++ examples/hello-world/package.json
@@ -14,6 +14,7 @@
     "react-dom": "16.4.1",
     "redux": "4.0.0",
     "resolve-scripts": "0.12.3",
+    "resolve-redux": "0.12.3",
     "react-helmet": "5.2.0",
     "react-bootstrap": "0.32.1"
   },
