+++ examples/with-postcss/package.json
@@ -18,9 +18,11 @@
   "dependencies": {
     "react": "16.4.1",
     "react-dom": "16.4.1",
+    "react-bootstrap": "0.32.1",
     "redux": "4.0.0",
     "react-helmet": "5.2.0",
-    "resolve-scripts": "0.12.3"
+    "resolve-scripts": "0.12.3",
+    "resolve-redux": "0.12.3"
   },
   "devDependencies": {
     "autoprefixer": "8.6.5",
