+++ examples/with-styled-components/package.json
@@ -12,9 +12,11 @@
   "dependencies": {
     "react": "16.4.1",
     "react-dom": "16.4.1",
+    "react-bootstrap": "0.32.1",
     "react-helmet": "5.2.0",
     "redux": "4.0.0",
     "resolve-scripts": "0.12.3",
+    "resolve-redux": "0.12.3",
     "styled-components": "3.3.3"
   },
   "devDependencies": {
