+++ packages/resolve-redux/src/connect_root_based_urls.js
@@ -1,58 +1,8 @@
-import React from 'react'
-
 import getRootBasedUrl from './get_root_based_url'
-import { Consumer } from './resolve_context'
-import * as validate from './validate'
+import createContextBasedConnector from './create_context_based_connector'
 
-const connectRootBasedUrls = propsList => Component => {
-  validate.arrayOfString(propsList, 'Props list')
-
-  const propsListSize = propsList.length
-
-  return class RootBasedComponent extends React.PureComponent {
-    functionAsChildComponent = ({ origin = '', rootPath = '' } = {}) => {
-      const props = this.props
-
-      const staticBasedProps = {}
-
-      for (
-        let propertyIndex = 0;
-        propertyIndex < propsListSize;
-        propertyIndex++
-      ) {
-        const propertyName = propsList[propertyIndex]
-        const propertyValue = props[propertyName]
-        if (Array.isArray(propertyValue)) {
-          const subProps = []
-          const subPropertySize = propertyValue.length
-          for (
-            let subPropertyIndex = 0;
-            subPropertyIndex < subPropertySize;
-            subPropertyIndex++
-          ) {
-            subProps[subPropertyIndex] = getRootBasedUrl(
-              origin,
-              rootPath,
-              propertyValue[subPropertyIndex]
-            )
-          }
-          staticBasedProps[propertyName] = subProps
-        } else {
-          staticBasedProps[propertyName] = getRootBasedUrl(
-            origin,
-            rootPath,
-            propertyValue
-          )
-        }
-      }
-
-      return <Component {...props} {...staticBasedProps} />
-    }
-
-    render() {
-      return <Consumer>{this.functionAsChildComponent}</Consumer>
-    }
-  }
-}
+const connectRootBasedUrls = createContextBasedConnector(
+  ({ origin, rootPath }, path) => getRootBasedUrl(origin, rootPath, path)
+)
 
 export default connectRootBasedUrls
